docker_requirements.txt
-i https://pypi.org/simple
aiohttp==3.8.6 ; python_version >= '3.6'
aiosignal==1.3.1 ; python_version >= '3.7'
altair==5.1.2 ; python_version >= '3.8'
async-timeout==4.0.3 ; python_version >= '3.7'
attrs==23.1.0 ; python_version >= '3.7'
blinker==1.6.3 ; python_version >= '3.7'
cachetools==5.3.2 ; python_version >= '3.7'
certifi==2023.7.22 ; python_version >= '3.6'
charset-normalizer==3.3.1 ; python_full_version >= '3.7.0'
click==8.1.7 ; python_version >= '3.7'
frozenlist==1.4.0 ; python_version >= '3.8'
gitdb==4.0.11 ; python_version >= '3.7'
gitpython==3.1.40 ; python_version >= '3.7'
idna==3.4 ; python_version >= '3.5'
importlib-metadata==6.8.0 ; python_version >= '3.8'
jinja2==3.1.2 ; python_version >= '3.7'
jsonschema==4.19.1 ; python_version >= '3.8'
jsonschema-specifications==2023.7.1 ; python_version >= '3.8'
markdown-it-py==3.0.0 ; python_version >= '3.8'
markupsafe==2.1.3 ; python_version >= '3.7'
mdurl==0.1.2 ; python_version >= '3.7'
multidict==6.0.4 ; python_version >= '3.7'
numpy==1.26.1 ; python_version < '3.13' and python_version >= '3.9'
openai==0.28.1
packaging==23.2 ; python_version >= '3.7'
pandas==2.1.2 ; python_version >= '3.9'
pillow==10.1.0 ; python_version >= '3.8'
protobuf==4.24.4 ; python_version >= '3.7'
pyarrow==13.0.0 ; python_version >= '3.8'
pydeck==0.8.1b0 ; python_version >= '3.7'
pygments==2.16.1 ; python_version >= '3.7'
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.0
pytz==2023.3.post1
referencing==0.30.2 ; python_version >= '3.8'
requests==2.31.0 ; python_version >= '3.7'
rich==13.6.0 ; python_full_version >= '3.7.0'
rpds-py==0.10.6 ; python_version >= '3.8'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
smmap==5.0.1 ; python_version >= '3.7'
streamlit==1.28.0
tenacity==8.2.3 ; python_version >= '3.7'
toml==0.10.2 ; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
toolz==0.12.0 ; python_version >= '3.5'
tornado==6.3.3 ; python_version >= '3.8'
tqdm==4.66.1 ; python_version >= '3.7'
typing-extensions==4.8.0 ; python_version >= '3.8'
tzdata==2023.3 ; python_version >= '2'
tzlocal==5.2 ; python_version >= '3.8'
urllib3==2.0.7 ; python_version >= '3.7'
validators==0.22.0 ; python_version >= '3.8'
yarl==1.9.2 ; python_version >= '3.7'
zipp==3.17.0 ; python_version >= '3.8'
